[기능추가] UI, 점수, 맵 전부 동시에 진행 중
- 게임모드 / 게임 스테이트 / 플레이어 스테이트 중심으로 작업 진행
- 플레이어 Kill, Death, Score 시스템
- 위 정보를 기반으로 정보를 표시하는 UI
- 메인 맵 편집

+++ Source/Explosion/Bomb/EPBombManager.h
@@ -32,9 +32,9 @@ private:
 	TArray<TObjectPtr<AEPBombBase>> BombList;
 
 	UPROPERTY(VisibleAnyWhere, Category = "Properties")
-	int8 MaxBombCount;
+	int32 MaxBombCount;
 
 	UPROPERTY(VisibleAnyWhere, Category = "Properties")
-	int8 CurrentBombCount;
+	int32 CurrentBombCount;
 
 };
